refactor: userClass -> useClassDetail, useClassList로 분리
