+++ src/components/common/BottomSheet/Body/TransferForm.jsx
@@ -1,22 +1,22 @@
 import styles from './BottomSheetForm.module.css';
 import { useBottomSheet } from '../../../../context/BottomSheetContext';
 import useInput from '../../../../hooks/useInput';
-import useClass from '../../../Main/hooks/useClass';
 import Input from '../../Input/Input';
 import Button from '../../Button/Button';
 import getTodayDate from '../../../../utils/getTodayDate';
+import useClassDetail from '../../../../hooks/useClassDetail';
 
 export default function TransferForm({ code, histories, text, tag, color }) {
   const { toggle, closeSheet } = useBottomSheet();
   const [info, handleChange] = useInput({ date: getTodayDate() });
-  const { isLoading, error, handleAddHistorySumbit } = useClass();
+  const { isLoading, error, handleAddHistorySumbit } = useClassDetail(code);
   const getMinDate = histories.find((history) => history.type === 'classModify');
 
   if (toggle !== tag) return;
   return (
     <form
       className={styles.form}
-      onSubmit={(e) => handleAddHistorySumbit(e, info, tag, code, getMinDate?.date, closeSheet)}
+      onSubmit={(e) => handleAddHistorySumbit(e, info, tag, getMinDate?.date, closeSheet)}
     >
       <Input type='number' name='price' text='금액' value={info} onChange={handleChange} required />
       <Input
